A first integration into stepper files
Compiles, more to check when printer arrives.
To check: All LIN_ADVANCE array are removed to a single value as there
should be never more than one stepper active. Does it work or makes
problems with multi-color printing?
Maybe todo for Multi Color: On a color change, reset all variables to 0
as if a new print would start. Pre-tension is not longer existing.